Clean up interfaces, naming conventions, and index file for exporting
