+++ client/src/interface/DialogItem.ts
@@ -1,0 +1,6 @@
+export interface IDialogItem {
+  title?: string;
+  content?: string;
+  icon?: string;
+  id?: string;
+}
