+++ client/src/interface/index.ts
@@ -1,0 +1,10 @@
+export type { IUser, ISearchUsersApiData } from './User';
+export type { IBoard } from './Board';
+export type { ICard } from './Card';
+export type { IColumn } from './Column';
+export type { IDialogItem } from './DialogItem';
+export type { IFetchOptions } from './FetchOptions';
+export type { IImageUploadData, IUploadOptions } from './File';
+export type { IAuthApiDataSuccess, IAuthApiData } from './AuthApiData';
+export type { IKanbanContext } from './KanbanContext';
+export type { IDialogItemContext } from './DialogItemContext';
